+++ card.hpp
@@ -10,10 +10,10 @@ public:
     object** objects;
 
     char* mapa;
-    size_t rows;
-    size_t length;
+    uint8_t rows;
+    uint8_t length;
 
-    za_mapo(size_t _rows, size_t _length, size_t pos_y, size_t pos_x);
+    za_mapo(uint8_t _rows, uint8_t _length, uint8_t pos_y, uint8_t pos_x);
     ~za_mapo();
 
     void init_palitra();
